+++ packages/polycli-runtime/src/index.js
@@ -8,4 +8,5 @@ export * from "./minimax.js";
 export * from "./opencode.js";
 export * from "./pi.js";
 export * from "./cmd.js";
+export * from "./agy.js";
 export * from "./registry.js";